+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.0.7 (not yet released)
 
-* While placing a terraforming project, resources that are currently in surplus are highlighted green.
+* While placing a terraforming project, resources that are currently in surplus are colored green in the production panel, and systems that only provide a single unit of that resource (and provide no other resources) are highlighted green on the map.
 
 ## 0.0.6 (2022-12-04)